==> code examples/old_script/2.5-Survey-twins.ipynb <==
%load_ext autoreload
%autoreload 2
%cd C:\MAD4AG
%matplotlib inline

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from tqdm.auto import tqdm
from collections import Counter


import warnings

warnings.filterwarnings('ignore')

df_survey_act= pd.read_pickle(f'./dbs/survey_act_weekday.pkl')
df_survey_dist= pd.read_pickle(f'./dbs/intermediate/df_survey_dist.pkl')



# create tuple showing all activities of an individual

df_survey_ppl= df_survey_act.groupby('sub_id')['purpose'].apply(tuple).reset_index(name='act_seq')

df_survey_ppl.act_seq.value_counts(normalize=False)

# find the unique activities from the activity tuple


def unique_activities(data):
    act_tuple = data['act_seq']
    data['act_uniq'] = tuple(Counter(act_tuple).keys())
    # this will take only unique numbers from the tuple
    return data


tqdm.pandas()
df_survey_ppl= df_survey_ppl.apply(lambda row: unique_activities(row) , axis=1 )

df_survey_ppl.act_uniq.value_counts()

# remove individuals having only home activity
df_survey_ppl = df_survey_ppl[df_survey_ppl.act_uniq != ('Home',)]

# bring the distance information

df_survey_ppl = pd.merge(df_survey_ppl, df_survey_dist, on='sub_id', how='left')

df_survey_ppl = df_survey_ppl[df_survey_ppl.avg_dist.notnull()]

df_survey_ppl['commute']= df_survey_ppl['commute'].astype(int)

df_survey_ppl['county']= df_survey_ppl['origin_main_deso'].str[:2]

df_survey_ppl['urban_density']= df_survey_ppl['origin_main_deso'].str[4]

df_survey_ppl.act_seq.value_counts(normalize=True)

df_survey_ppl.to_pickle(r'.\dbs\intermediate\df_survey_twins.pkl')
